Fix a bug preventing enum_by working after errno
In case a former call to the Windows API created an error
condition (i.e., setting an error on SetErrorEx()), then
enum_by and enum_by_until would incorrectly always
return an error as a result due to the former error not
being cleaned. .Since the contract on safe Rust functions
is usually to return their own error (and not carry other
former errors to the user) it makes sense that enum_by
would clean the former error with SetErrorEx(0, 0) before
operating.

This commit contributes the fix and also a test proving
the failure to guard against a regression.